add CRUD for products

+++ src/components/RecordsEntity/index.js
@@ -4,19 +4,21 @@ import './index.css'
 
 export class RecordEntity extends Component {
   handleDeleteRecord = () => {
-    const { id, onDelete } = this.props
-    onDelete(id)
+    const { id, onDelete, getProducts } = this.props
+    if (window.confirm('Вы действительно хотите удалить запись?')) {
+      onDelete(id)
+    }
   }
 
   render () {
     const { recordEntity } = this.props
-    const { name, description, initialPrice, price } = recordEntity
+    const { name, flavor, initial_price, full_price } = recordEntity
     return (
       <tr className='juice-tracker-RecordsEntity-item'>
         <td>{name}</td>
-        <td>{description}</td>
-        <td>{initialPrice}</td>
-        <td>{price}</td>
+        <td>{flavor}</td>
+        <td>{initial_price}</td>
+        <td>{full_price}</td>
         <td>
           <Button
             className='juice-tracker-RecordsEntity-btn'
